Improve the performance of attribute parsing for .b3dm
The previous implementation would use a lot of string allocations which
would cause the garbage collector to kick in often and that was a major
fit for performance.

This patch fixes the issue by using the built-in TextDecoder class, but
it might break compatibility with old browsers (e.g. Internet Explorer).