[types] - three types from versions after they split them to a different repo are not compatible with older versions
This doesn't make much difference in practice for usage, but it fixes issues where the three types are not compatible when the application using this package is on a newer version of three.

(I chose 128 just because that is the version that we've been using for quite a while with no problems, it may also simply be able to use a newer version if desired? We've also tested against three 130, though not as thoroughly as 128.)