fix: improve content-uri handling
The old implementation didn't handle absolute paths properly (#255), so
this is an attempt at solving that issue.

It turned out that the `urlJoin` utility-function was only used in a
single place, where it could be replaced with the URL constructor, so I
removed both the utility-function as well as its tests.

That left only a single call to `path.basename` remaining as the reason
to have the `path-browserify` dependency, so I replaced that as well and
removed the dependency.